+++ engine/src/io/data_parser/metadata/orc_metadata.cpp
@@ -4,45 +4,46 @@
 #include "orc_metadata.h"
 //#include "ExceptionHandling/BlazingThread.h"
 #include "utilities/CommonOperations.h"
-//#include <cudf/column/column_factories.hpp>
 
-
-// TODO: this function is the same as the parquet_metadata::makeMetadataTable
-// we just want one of them
-std::unique_ptr<ral::frame::BlazingTable> makeMetadataORCTable(std::vector<std::string> col_names) {
-	const int ncols = col_names.size();
-	std::vector<std::string> metadata_col_names;
-	// + 2: due to file_handle_index and stripe_index
-	metadata_col_names.resize(ncols * 2 + 2);
-
-	int metadata_col_index = -1;
-	for (int colIndex = 0; colIndex < ncols; ++colIndex){
-		std::string col_name = col_names[colIndex];
-		std::string col_name_min = "min_" + std::to_string(colIndex) + "_" + col_name;
-		std::string col_name_max = "max_" + std::to_string(colIndex)  + "_" + col_name;
-
-		metadata_col_names[++metadata_col_index] = col_name_min;
-		metadata_col_names[++metadata_col_index] = col_name_max;
-	}
-
-	metadata_col_names[++metadata_col_index] = "file_handle_index";
-	metadata_col_names[++metadata_col_index] = "row_group_index"; // stripe_index
-
-	std::vector<std::unique_ptr<cudf::column>> minmax_metadata_gdf_table;
-	minmax_metadata_gdf_table.resize(metadata_col_names.size());
-	for (std::size_t i = 0; i < metadata_col_names.size(); ++i) {
-		std::vector<int32_t> temp{(int32_t)-1};
-		std::unique_ptr<cudf::column> expected_col = ral::utilities::vector_to_column(temp, cudf::data_type(cudf::type_id::INT32));
-		minmax_metadata_gdf_table[i] = std::move(expected_col);
-	}
-
-	auto cudf_metadata_table = std::make_unique<cudf::table>(std::move(minmax_metadata_gdf_table));
-	auto metadata_table = std::make_unique<ral::frame::BlazingTable>(std::move(cudf_metadata_table), metadata_col_names);
-
-	return metadata_table;
+std::basic_string<char> get_typed_vector_str_content(cudf::type_id dtype, std::vector<std::string> & vector) {
+	std::basic_string<char> output = std::basic_string<char>((char *)vector.data(), vector.size() * sizeof(char));
+	return output;
 }
 
-cudf::type_id to_dtype(cudf::io::statistics_type stat_type) {
+std::vector<int64_t> get_all_values_in_the_same_col( 
+	std::vector<std::vector<std::int64_t>> & min_max,
+	std::size_t index) {
+	std::vector<std::int64_t> output_v;
+	for (std::size_t col_index; col_index < min_max.size(); ++col_index) {
+		output_v.push_back(min_max[col_index][index]);
+	}
+
+	return output_v;
+}
+
+std::vector<std::string> get_all_str_values_in_the_same_col(
+	std::vector<std::vector<std::string>> & minmax_string,
+	std::size_t index) {
+	std::vector<std::string> output_v;
+	for (std::size_t col_index; col_index < minmax_string.size(); ++col_index) {
+		output_v.push_back(minmax_string[col_index][index]);
+	}
+
+	return output_v;
+}
+
+void set_min_max_string(
+	std::vector<std::string> & minmax_string_metadata_table,
+	cudf::io::column_statistics & statistic, int col_index) {
+
+	auto type_stat = statistic.type_specific_stats<cudf::io::string_statistics>();
+	std::string min = *type_stat->minimum();
+	std::string max = *type_stat->maximum();
+	minmax_string_metadata_table[col_index] = min;
+	minmax_string_metadata_table[col_index + 1] = max;
+}
+
+cudf::type_id statistic_to_dtype(cudf::io::statistics_type stat_type) {
 	if (stat_type == cudf::io::statistics_type::INT) {
 		return cudf::type_id::INT32;
 	} else if (stat_type == cudf::io::statistics_type::STRING) {
@@ -56,17 +57,6 @@ cudf::type_id to_dtype(cudf::io::statistics_type stat_type) {
 	} else {
 		return cudf::type_id::EMPTY;
 	}
-}
-
-void set_min_max_string(
-	std::vector<std::string> & minmax_string_metadata_table,
-	cudf::io::column_statistics & statistic, int col_index) {
-
-	auto type_stat = statistic.type_specific_stats<cudf::io::string_statistics>();
-	std::string min = *type_stat->minimum();
-	std::string max = *type_stat->maximum();
-	minmax_string_metadata_table[col_index] = min;
-	minmax_string_metadata_table[col_index + 1] = max;
 }
 
 void set_min_max(
@@ -102,92 +92,7 @@ void set_min_max(
 	}
 }
 
-std::basic_string<char> get_typed_vector_str_content_(cudf::type_id dtype, std::vector<std::string> & vector) {
-	std::basic_string<char> output = std::basic_string<char>((char *)vector.data(), vector.size() * sizeof(char));
-	return output;
-}
-
-std::basic_string<char> get_typed_vector_content_(cudf::type_id dtype, std::vector<int64_t> & vector) {
-  	std::basic_string<char> output;
-  	switch (dtype) {
-	case cudf::type_id::INT8:{
-			std::vector<char> typed_v(vector.begin(), vector.end());
-			output = std::basic_string<char>((char *)typed_v.data(), typed_v.size() * sizeof(char));
-			break;
-		}
-	case cudf::type_id::INT16: {
-		std::vector<int16_t> typed_v(vector.begin(), vector.end());
-		output = std::basic_string<char>((char *)typed_v.data(), typed_v.size() * sizeof(int16_t));
-		break;
-	}
-	case cudf::type_id::INT32:{
-		std::vector<int32_t> typed_v(vector.begin(), vector.end());
-		output = std::basic_string<char>((char *)typed_v.data(), typed_v.size() * sizeof(int32_t));
-		break;
-	}
-	case cudf::type_id::INT64: {
-		output = std::basic_string<char>((char *)vector.data(), vector.size() * sizeof(int64_t));
-		break;
-	}
-	case cudf::type_id::FLOAT64: {
-		double* casted_metadata = reinterpret_cast<double*>(&(vector[0]));
-		output = std::basic_string<char>((char *)casted_metadata, vector.size() * sizeof(double));
-		break;
-	}
-	case cudf::type_id::BOOL8: {
-		std::vector<int8_t> typed_v(vector.begin(), vector.end());
-		output = std::basic_string<char>((char *)typed_v.data(), typed_v.size() * sizeof(int8_t));
-		break;
-	}
-	case cudf::type_id::TIMESTAMP_SECONDS: {
-		output = std::basic_string<char>((char *)vector.data(), vector.size() * sizeof(int64_t));
-		break;
-	}
-	default: {
-		// default return type since we're throwing an exception.
-		std::cerr << "Invalid gdf_dtype in create_host_column" << std::endl;
-		throw std::runtime_error("Invalid gdf_dtype in create_host_column");
-	}
-  }
-  return output;
-}
-
-std::unique_ptr<cudf::column> make_cudf_column_from_orc(cudf::data_type dtype, std::basic_string<char> &vector, unsigned long column_size) {
-	size_t width_per_value = cudf::size_of(dtype); // TODO: issue when column is string
-	if (vector.size() != 0) {
-		auto buffer_size = width_per_value * column_size;
-		rmm::device_buffer gpu_buffer(vector.data(), buffer_size);
-		return std::make_unique<cudf::column>(dtype, column_size, std::move(gpu_buffer));
-	} else {
-		auto buffer_size = width_per_value * column_size;
-		rmm::device_buffer gpu_buffer(buffer_size);
-		return std::make_unique<cudf::column>(dtype, column_size, buffer_size);
-	}
-}
-
-std::vector<int64_t> get_all_values_in_the_same_col( 
-	std::vector<std::vector<std::int64_t>> & min_max,
-	std::size_t index) {
-	std::vector<std::int64_t> output_v;
-	for (std::size_t col_index; col_index < min_max.size(); ++col_index) {
-		output_v.push_back(min_max[col_index][index]);
-	}
-
-	return output_v;
-}
-
-std::vector<std::string> get_all_str_values_in_the_same_col(
-	std::vector<std::vector<std::string>> & minmax_string,
-	std::size_t index) {
-	std::vector<std::string> output_v;
-	for (std::size_t col_index; col_index < minmax_string.size(); ++col_index) {
-		output_v.push_back(minmax_string[col_index][index]);
-	}
-
-	return output_v;
-}
-
-std::unique_ptr<ral::frame::BlazingTable> get_minmax_orc_metadata(
+std::unique_ptr<ral::frame::BlazingTable> get_minmax_metadata(
     std::vector<cudf::io::parsed_orc_statistics> & orc_statistics,
     size_t total_stripes, int metadata_offset) {
 
@@ -218,7 +123,7 @@ std::unique_ptr<ral::frame::BlazingTable> get_minmax_orc_metadata(
 		std::vector<std::string> col_names = orc_statistics[0].column_names;
 		col_names.erase(col_names.begin());
 
-		return makeMetadataORCTable(col_names);
+		return makeMetadataFromCols(col_names);
 	}
 
 	// Getting metadata and stats from the whole orc file
@@ -230,7 +135,7 @@ std::unique_ptr<ral::frame::BlazingTable> get_minmax_orc_metadata(
 
 	if (num_stripes > 0) {
 		for (std::size_t colIndex = 0; colIndex < file_metadata.size(); colIndex++) {
-			cudf::data_type dtype = cudf::data_type(to_dtype(file_metadata[colIndex].type())) ;
+			cudf::data_type dtype = cudf::data_type(statistic_to_dtype(file_metadata[colIndex].type())) ;
 			if (file_metadata[colIndex].type() != cudf::io::statistics_type::NONE) {
 				std::string col_name_min = "min_" + std::to_string(colIndex) + "_" + col_names[colIndex];
 				metadata_names.push_back(col_name_min);
@@ -317,16 +222,16 @@ std::unique_ptr<ral::frame::BlazingTable> get_minmax_orc_metadata(
 		if (dtype == cudf::data_type{cudf::type_id::STRING}) {
 			std::vector<std::string> vector = get_all_str_values_in_the_same_col(minmax_string_metadata, string_count);
 			string_count++;
-			content = get_typed_vector_str_content_(dtype.id(), vector); // TODO review if this is handled correctecly
+			content = get_typed_vector_str_content(dtype.id(), vector); // TODO review if this is handled correctecly
 			dtype = cudf::data_type{cudf::type_id::INT32}; // TODO: avoid this ? Issue when no set up to int32
 		} else {
 			std::vector<int64_t> vector = get_all_values_in_the_same_col(minmax_metadata, not_string_count);
 			not_string_count++;
-			content = get_typed_vector_content_(dtype.id(), vector);
+			content = get_typed_vector_content(dtype.id(), vector);
 			//std::unique_ptr<cudf::column> expected_col2 = ral::utilities::vector_to_column(vector, dtype);
 			//minmax_metadata_gdf_table[index] = std::move(expected_col2);
 		}
-		minmax_metadata_gdf_table[index] = make_cudf_column_from_orc(dtype, content, total_stripes);
+		minmax_metadata_gdf_table[index] = make_cudf_column_from(dtype, content, total_stripes);
 	}
 
 	auto table = std::make_unique<cudf::table>(std::move(minmax_metadata_gdf_table));
