+++ engine/src/utilities/CommonOperations.cpp
@@ -121,7 +121,7 @@ std::unique_ptr<ral::frame::BlazingTable> create_empty_table(const std::vector<s
 	std::vector<std::unique_ptr<cudf::column>> columns(column_indices.size());
 
 	for (auto idx : column_indices) {
-		columns[idx] =  make_empty_column(dtypes[idx]);
+		columns[idx] = cudf::make_empty_column(dtypes[idx]);
 	}
 	auto table = std::make_unique<cudf::table>(std::move(columns));
 	return std::make_unique<ral::frame::BlazingTable>(std::move(table), column_names);
@@ -130,7 +130,7 @@ std::unique_ptr<ral::frame::BlazingTable> create_empty_table(const std::vector<s
 std::unique_ptr<cudf::table> create_empty_table(const std::vector<cudf::type_id> &dtypes) {
 	std::vector<std::unique_ptr<cudf::column>> columns(dtypes.size());
 	for (size_t idx =0; idx < dtypes.size(); idx++) {
-		columns[idx] =  make_empty_column(cudf::data_type(dtypes[idx]));
+		columns[idx] = cudf::make_empty_column(cudf::data_type(dtypes[idx]));
 	}
 	return std::make_unique<cudf::table>(std::move(columns));
 }
